Allow dev to modify the return key type for each field and the submit action
This will allow you to have a name and billing zip code field present on the page and on submit from the name field go to the zip code field for a better user experience. Currently, the form blurs and you have to click the billing zip code.

minor fix

Allow the developer to pass down additional input props

This should probably change more of the code than the above.

Generalizing so the developer can pass whatever to the inputs

in the form of:
```javascript
additionalInputProps = {
  name: { returnKeyType: 'next' },
};
```

Same thing on the lite inputs

clean up lint error

fix linting errors

clearning up prop names

Update Readme for the additionalInputsProps

fixing typo

fixing propTypes